Update to Tuxpaint 0.9.24
Tuxpaint 0.9.24 uses xdg-icon-resource and xdg-desktop-menu to install
icons and desktop files. Unfortunately flatpak-xdg-utils does not
provide these.  https://github.com/flatpak/flatpak-xdg-utils/issues/35

I started by building the real xdg-utils, patching it to avoid a
transitive dependency on Links or W3M. However because the build runs as
a non-root user, these tools default to installing into $HOME.

I then tried reimplementing simple versions of these tools. It worked,
but we still need to patch the Makefile for the SVG icon and to remove
the call to update-desktop-database, so, we may as well just do that as
before.

The other two patches needed very minor refreshment.