Add flatpak-external-data-checker goop
This fetches the latest tuxpaint version from
https://release-monitoring.org/. Although tuxpaint-stamps is on there,
the version number can't be simply interpolated into the URL, so I
haven't added that integration.

As another hiccup, the checker assumes that the last module is the app
itself; since tuxpaint-stamps comes after tuxpaint, updates to the
latter won't update the app version in appdata. Never mind!

(I tried reversing the modules and it's not as simple as that.)